# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Expedisoft",
        "description": "Documentação da API de integração do sistema Expedisoft",
        "contact": {
            "email": "suporte@expedisoft.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Servidor API Local"
        }
    ],
    "paths": {
...
        "/api/integration/user": {
            "post": {
                "tags": [
                    "Integration"
                ],
                "summary": "Create or update a user",
                "operationId": "12f0b4b8d7ef790ad36ded1ac9b0c059",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "source_system",
                                    "user"
                                ],
                                "properties": {
                                    "source_system": {
                                        "type": "string",
                                        "example": "SAP"
                                    },
                                    "user": {
                                        "required": [
                                            "external_id",
                                            "name",
                                            "email"
                                        ],
                                        "properties": {
                                            "external_id": {
                                                "type": "string",
                                                "example": "12345"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "teste@email.com"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Payload received and queued for processing"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/order/{orderId}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Busca uma ordem específica ou lista todas as ordens",
                "description": "Retorna os detalhes de uma ordem se o ID for passado. Se não, retorna uma lista paginada com todas as ordens.",
                "operationId": "0a8bf88e870ceb6678a8e2f55eddef75",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "UUID da ordem de carregamento",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação realizada com sucesso"
                    },
                    "404": {
                        "description": "Ordem de carregamento não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/order/my-orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Lista as ordens atreladas ao usuário logado",
                "description": "Retorna uma lista paginada de ordens criadas por ou designadas para o operador atual.",
                "operationId": "2e42fbc21a9ff6ac6b42f8b2eb70d522",
                "responses": {
                    "200": {
                        "description": "Operação realizada com sucesso"
                    },
                    "404": {
                        "description": "Nenhuma ordem encontrada para o usuário"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/order/{orderId}/start-order": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Inicia o processo de uma ordem de carregamento",
                "operationId": "9053b139fb6c7348819f38f72e45221a",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "UUID da ordem a ser iniciada",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ordem iniciada com sucesso"
                    },
                    "400": {
                        "description": "A ordem não está no status 'scheduled'"
                    },
                    "403": {
                        "description": "Você não tem permissão para iniciar esta ordem"
                    },
                    "404": {
                        "description": "Ordem de carregamento não encontrada"
                    },
                    "500": {
                        "description": "Erro interno no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/order/{orderId}/finish-order": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Finaliza uma ordem de carregamento",
                "operationId": "f39fcf883f221f93775b1f4e49d7d90d",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "UUID da ordem a ser finalizada",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ordem finalizada com sucesso"
                    },
                    "400": {
                        "description": "A ordem não está no status 'in_progress'"
                    },
                    "403": {
                        "description": "Você não tem permissão para finalizar esta ordem"
                    },
                    "404": {
                        "description": "Ordem de carregamento não encontrada"
                    },
                    "500": {
                        "description": "Erro interno no servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}